% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{dms_EdiOrderConfirm_delete}
\alias{dms_EdiOrderConfirm_delete}
\title{EDI数据中台销售订单确认删除}
\usage{
dms_EdiOrderConfirm_delete(dms_token, FBillNo, FDate)
}
\arguments{
\item{FDate}{}
}
\value{
两个数的和
}
\description{
EDI数据中台销售订单确认删除
}
\examples{
dms_EdiOrderConfirm_delete()
}
