% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderConfirm_delete}
\alias{EdiOrderConfirm_delete}
\title{EDI销售订单确认删除}
\usage{
EdiOrderConfirm_delete(erp_token, FMessageNumber, FDate)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单确认删除
}
\examples{
EdiOrderConfirm_delete()
}
