% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderConfirm_sync}
\alias{EdiOrderConfirm_sync}
\title{EDI销售订单同步}
\usage{
EdiOrderConfirm_sync(erp_token, FMessageNumber)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单同步
}
\examples{
EdiOrderConfirm_sync()
}
