% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderConfirm_update}
\alias{EdiOrderConfirm_update}
\title{EDI销售订单更新}
\usage{
EdiOrderConfirm_update(
  erp_token,
  FMessageNumber,
  FLineItemNumber,
  FCommittedQuantity,
  FCommittedQuantityUOM,
  FCommittedQuantityConfirmedDate,
  FErpDeliveryDate
)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单更新
}
\examples{
EdiOrderConfirm_update()
}
