% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderFIsDo_update}
\alias{EdiOrderFIsDo_update}
\title{EDI销售订单状态更新}
\usage{
EdiOrderFIsDo_update(erp_token, FMessageNumber)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单状态更新
}
\examples{
EdiOrderFIsDo_update()
}
