% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderHeader_view}
\alias{EdiOrderHeader_view}
\title{EDI销售订单表头查询}
\usage{
EdiOrderHeader_view(erp_token, FMessageNumber)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单表头查询
}
\examples{
EdiOrderHeader_view()
}
