% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderItem_view}
\alias{EdiOrderItem_view}
\title{EDI销售订单明细查询}
\usage{
EdiOrderItem_view(erp_token, FMessageNumber)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单明细查询
}
\examples{
EdiOrderItem_view()
}
