% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderConfirm_view}
\alias{EdiOrderConfirm_view}
\title{EDI销售订单查询}
\usage{
EdiOrderConfirm_view(erp_token, FMessageNumber, FLineItemNumber)
}
\arguments{
\item{erp_token}{}
}
\value{
两个数的和
}
\description{
EDI销售订单查询
}
\examples{
EdiOrderConfirm_view()
}
